+++ ChangeLog.md
@@ -8,6 +8,7 @@
 	* add: starter-tomcat by @ben
 	* add: starter-beetlsql 来自beetl的SQL解决方案
 	* add: starter-sharding-jdbc 分库分表
+	* add: starter-thymeleaf 模板引擎
 	* fix: jetty扫描websocket的endpoint有问题
 
 ## 2.0-RC "属于"